backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==1.1.0
google-genai==0.3.0
python-dotenv==1.0.0
reportlab==4.0.7
PyJWT==2.8.0
requests==2.31.0
psycopg
psycopg[binary]
bcrypt
openai
gunicorn